+++ examples/live/ftx_ema_cross.py
@@ -75,7 +75,7 @@ strat_config = EMACrossConfig(
     bar_type="ETH-PERP.FTX-1-MINUTE-LAST-INTERNAL",
     fast_ema_period=10,
     slow_ema_period=20,
-    trade_size=Decimal("0.01"),
+    trade_size=Decimal("0.010"),
     order_id_tag="001",
 )
 # Instantiate your strategy