+++ nautilus_trader/adapters/betfair/client/core.py
@@ -17,7 +17,7 @@ import asyncio
 import datetime
 import pathlib
 import ssl
-from typing import Dict, List, Optional
+from typing import Optional
 
 import msgspec
 from aiohttp import ClientResponse
@@ -91,8 +91,8 @@ class BetfairClient(HttpClient):
         return await super().request(method=method, url=url, **kwargs)
 
     async def rpc_post(
-        self, url, method, params: Optional[Dict] = None, data: Optional[Dict] = None
-    ) -> Dict:
+        self, url, method, params: Optional[dict] = None, data: Optional[dict] = None
+    ) -> dict:
         data = {**self.JSON_RPC_DEFAULTS, "method": method, **(data or {}), "params": params or {}}
         try:
             resp = await self.request(method="POST", url=url, headers=self.headers, json=data)
@@ -147,7 +147,7 @@ class BetfairClient(HttpClient):
     async def list_market_catalogue(
         self,
         filter_: dict,
-        market_projection: List[MarketProjection] = None,
+        market_projection: list[MarketProjection] = None,
         sort: str = None,
         max_results: int = 1000,
         locale: str = None,
@@ -165,20 +165,25 @@ class BetfairClient(HttpClient):
         if "sort" in params:
             assert isinstance(sort, MarketSort)
         resp = await self.rpc_post(
-            url=self.BETTING_URL, method="SportsAPING/v1.0/listMarketCatalogue", params=params
+            url=self.BETTING_URL,
+            method="SportsAPING/v1.0/listMarketCatalogue",
+            params=params,
         )
         return resp
 
     async def get_account_details(self):
         resp = await self.rpc_post(
-            url=self.ACCOUNT_URL, method="AccountAPING/v1.0/getAccountDetails"
+            url=self.ACCOUNT_URL,
+            method="AccountAPING/v1.0/getAccountDetails",
         )
         return resp
 
     async def get_account_funds(self, wallet: Optional[str] = None):
         params = parse_params(**locals())
         resp = await self.rpc_post(
-            url=self.ACCOUNT_URL, method="AccountAPING/v1.0/getAccountFunds", params=params
+            url=self.ACCOUNT_URL,
+            method="AccountAPING/v1.0/getAccountFunds",
+            params=params,
         )
         return resp
 
@@ -187,7 +192,7 @@ class BetfairClient(HttpClient):
         market_id: str,
         instructions: list,
         customer_ref: str = None,
-        market_version: Optional[Dict] = None,
+        market_version: Optional[dict] = None,
         customer_strategy_ref: str = None,
     ):
         """
@@ -195,7 +200,9 @@ class BetfairClient(HttpClient):
         """
         params = parse_params(**locals())
         resp = await self.rpc_post(
-            url=self.BETTING_URL, method="SportsAPING/v1.0/placeOrders", params=params
+            url=self.BETTING_URL,
+            method="SportsAPING/v1.0/placeOrders",
+            params=params,
         )
         return resp
 
@@ -204,11 +211,13 @@ class BetfairClient(HttpClient):
         market_id: str = None,
         instructions: list = None,
         customer_ref: str = None,
-        market_version: Optional[Dict] = None,
+        market_version: Optional[dict] = None,
     ):
         params = parse_params(**locals())
         resp = await self.rpc_post(
-            url=self.BETTING_URL, method="SportsAPING/v1.0/replaceOrders", params=params
+            url=self.BETTING_URL,
+            method="SportsAPING/v1.0/replaceOrders",
+            params=params,
         )
         return resp
 
@@ -220,7 +229,9 @@ class BetfairClient(HttpClient):
     ):
         params = parse_params(**locals())
         resp = await self.rpc_post(
-            url=self.BETTING_URL, method="SportsAPING/v1.0/cancelOrders", params=params
+            url=self.BETTING_URL,
+            method="SportsAPING/v1.0/cancelOrders",
+            params=params,
         )
         return resp
 
@@ -238,7 +249,7 @@ class BetfairClient(HttpClient):
         from_record: int = None,
         record_count: int = None,
         include_item_description: bool = None,
-    ) -> List[Dict]:
+    ) -> list[dict]:
         params = parse_params(**locals())
         current_orders = []
         more_available = True
@@ -246,7 +257,9 @@ class BetfairClient(HttpClient):
         while more_available:
             params["fromRecord"] = index
             resp = await self.rpc_post(
-                url=self.BETTING_URL, method="SportsAPING/v1.0/listCurrentOrders", params=params
+                url=self.BETTING_URL,
+                method="SportsAPING/v1.0/listCurrentOrders",
+                params=params,
             )
             order_chunk = resp["currentOrders"]
             current_orders.extend(order_chunk)
@@ -272,7 +285,7 @@ class BetfairClient(HttpClient):
         locale: str = None,
         from_record: int = None,
         record_count: int = None,
-    ) -> List[Dict]:
+    ) -> list[dict]:
         params = parse_params(**locals())
         cleared_orders = []
         more_available = True
@@ -282,7 +295,9 @@ class BetfairClient(HttpClient):
             if settled_date_from or settled_date_to:
                 params["settledDateRange"] = {"from": settled_date_from, "to": settled_date_to}
             resp = await self.rpc_post(
-                url=self.BETTING_URL, method="SportsAPING/v1.0/listClearedOrders", params=params
+                url=self.BETTING_URL,
+                method="SportsAPING/v1.0/listClearedOrders",
+                params=params,
             )
             order_chunk = resp["clearedOrders"]
             cleared_orders.extend(order_chunk)